user-to-review now one-to-many; admin can view and delete reviews

+++ server/src/domain/review.rs
@@ -4,7 +4,7 @@ use uuid::Uuid;
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Review {
     pub review_id:  Uuid,
-    pub contact_id: i64,
+    pub contact_id: i32,
     pub user_id:    Uuid,
     pub title:      Option<String>,
     pub body:       Option<String>,
@@ -23,7 +23,7 @@ pub struct ReviewResponse {
 #[derive(Debug, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ContactReview {
-    pub contact_id: i64,
+    pub contact_id: i32,
     pub rating:     f32,
     pub reviews:    Vec<ReviewResponse>
 }
